extract TuningSelectorCommon from TuningColumn
