+++ src/modules/fretboard/components/TuningColumn.tsx
@@ -1,17 +1,12 @@
-import { ChangeEventHandler } from 'react';
-
 import { changeTuning } from '@modules/tablatureStore/actions/changeTuning';
 import { useTablatureStore } from '@modules/tablatureStore/useTablatureStore';
 
 import styles from './Fretboard.module.scss';
 import TuningSelector from './TuningSelector';
+import TuningSelectorCommon from './TuningSelectorCommon';
 
 const TuningColumn = () => {
 	const tuning = useTablatureStore().tuning;
-	const instrument = useTablatureStore().instrument;
-
-	const onCommonTuningChange: ChangeEventHandler<HTMLSelectElement> = (e) =>
-		changeTuning(instrument.commonTunings[e.target.value]);
 
 	const onTuningSelectorChange = (stringIndex: number, tuningValue: number) => {
 		const newTuning = [...tuning];
@@ -21,13 +16,7 @@ const TuningColumn = () => {
 
 	return (
 		<div className={styles['tuning-column']}>
-			<select defaultValue={instrument.defaultTuningName} onChange={onCommonTuningChange}>
-				{Object.keys(instrument.commonTunings).map((tuningName) => (
-					<option key={tuningName} value={tuningName}>
-						{tuningName === instrument.defaultTuningName ? `${tuningName} (Default)` : tuningName}
-					</option>
-				))}
-			</select>
+			<TuningSelectorCommon />
 			{tuning.map((tuningValue, stringIndex) => (
 				<TuningSelector
 					key={stringIndex}
